requirements.txt
asgiref==3.5.2
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
Pillow==9.2.0
pytz==2022.4
sqlparse==0.4.3
